+++ src/ pages/ProductList/ProductList.scss
@@ -6,44 +6,35 @@
 }
 
 .productTitle {
+  text-align: center;
+  width: 75vw;
+  margin: auto;
+  display: grid;
+  margin-bottom: 15px;
+
   h2 {
+    color: #333333;
+    margin-top: 20px;
     margin-bottom: 20px;
     font-size: 1.8rem;
   }
-  text-align: center;
-  width: 85vw;
-  margin: 0 auto;
-  display: grid;
-  margin-bottom: 15px;
+
+  p {
+    margin-bottom: 40px;
+  }
 
   ul {
     padding: 20px 0 20px 0;
-    border: 1px solid rgb(135, 135, 135);
+    border: 1px solid #e2e2e2;
     li {
+      color: #333333;
       display: inline-block;
-      margin: 0 100px;
-      font-size: 1.3rem;
-      cursor: pointer;
-    }
-  }
-}
-
-.soltMenu {
-  width: 85vw;
-  margin: 0 auto;
-  margin-bottom: 15px;
-  display: grid;
-  ul {
-    li {
-      display: inline-block;
-      margin: 0 10px;
-      float: right;
-      color: rgb(135, 135, 135);
+      margin: 0 120px;
       font-size: 1rem;
       cursor: pointer;
 
       &:hover {
-        color: black;
+        color: #5e0080;
       }
     }
   }
