+++ src/ pages/ProductList/components/MenuList/MenuList.scss
@@ -1,7 +1,9 @@
 .menuList {
-  width: 85vw;
+  width: 75vw;
   margin: 0 auto;
   display: grid;
   grid-template-columns: 1fr 1fr 1fr;
-  grid-gap: 20px;
+  grid-gap: 25px;
+  align-items: center;
+  justify-items: center;
 }
