+++ src/ pages/ProductList/components/SortMenu/SortMenu.scss
@@ -1,0 +1,21 @@
+.soltMenu {
+  width: 75vw;
+  margin: 0 auto;
+  margin-bottom: 15px;
+  display: grid;
+
+  ul {
+    li {
+      display: inline-block;
+      margin: 0 10px;
+      float: right;
+      color: rgb(135, 135, 135);
+      font-size: 0.8rem;
+      cursor: pointer;
+
+      &:hover {
+        color: black;
+      }
+    }
+  }
+}
