+++ src/ pages/ProductList/components/cartBtn/CartBtn.js
@@ -1,0 +1,12 @@
+import React from 'react';
+// import './cartBtn.scss';
+
+function CartBtn() {
+  return (
+    <div className="cartBtn">
+      <button class="btn1">버튼1</button>
+    </div>
+  );
+}
+
+export default CartBtn;
