+++ src/ pages/ProductList/components/cartBtn/CartBtn.scss
@@ -1,0 +1,6 @@
+// .cartBtn {
+//   button {
+//     width: 45px;
+//     height: 45px;
+//   }
+// }
